Extract gong synthesis into dedicated module
